+++ examples/scoring/data/homotrimer.pdb
