+++ examples/scoring/data/ligand-prodrg.param
