+++ examples/scoring/data/ligand-prodrg.top
