+++ examples/scoring/data/prot-lig.pdb
