+++ examples/scoring/data/prot-prot.pdb
