+++ examples/scoring/prodigy-protein-ligand-test.cfg
@@ -1,4 +1,4 @@
-run_dir = "run-prodigy-prot-lig"
+run_dir = "run-prodigy-lig"
 molecules = "data/prot-lig.pdb"
 
 [topoaa]
@@ -6,7 +6,7 @@ ligand_param_fname = "data/ligand-prodrg.param"
 ligand_top_fname = "data/ligand-prodrg.top"
 delenph = false
 
-[prodigy]
-scoring_mode = "ligand"
+[prodigyligand]
+receptor_chain = "A"
+ligand_chain = "B"
 ligand_resname = "G39"
-
