+++ examples/scoring/prodigy-protein-protein-test.cfg
@@ -4,7 +4,8 @@ molecules = [
     "data/prot-prot.pdb",
     "data/homotrimer.pdb",
     ]
-postprocess=false
-clean=false
+
 [topoaa]
-[prodigy]
+
+[prodigyprotein]
+
