+++ src/haddock/modules/scoring/prodigyligand/defaults.yaml
